requirements.txt
tensorflow
keras
segmentation-models
efficientnet
opencv-python
Pillow
numpy
matplotlib
gradio
